refactor(kit)!: relocate json-render from devframe to devtools-kit
JSON-render is a hub-layer concern: the spec is rendered by the
Vite DevTools component catalog, the renderer handle is only
consumable via a `DevToolsViewJsonRender` dock entry (kit type),
and every example pairs `createJsonRenderer` with
`ctx.docks.register({ type: 'json-render', ui })`. Move types,
factory, and `defineJsonRenderSpec` to `@vitejs/devtools-kit` so
the API lives in the package that owns the runtime contract;
`DevToolsNodeContext` is now framework-neutral as intended.

BREAKING: `ctx.createJsonRenderer` is no longer on
`DevToolsNodeContext`; use `KitNodeContext` (passed to
`Plugin.devtools.setup` and `createPluginFromDevframe`'s
`options.setup`). All in-repo consumers already run under kit.
