+++ devframe/packages/devframe/src/define.ts
@@ -1,8 +1,4 @@
-import type { DevToolsNodeContext, JsonRenderSpec } from 'devframe/types'
+import type { DevToolsNodeContext } from 'devframe/types'
 import { createDefineWrapperWithContext } from 'devframe/rpc'
 
 export const defineRpcFunction = createDefineWrapperWithContext<DevToolsNodeContext>()
-
-export function defineJsonRenderSpec(spec: JsonRenderSpec): JsonRenderSpec {
-  return spec
-}
