+++ devframe/packages/devframe/src/types/index.ts
@@ -4,7 +4,6 @@ export * from './devtool'
 export * from './diagnostics'
 export * from './events'
 export * from './host'
-export * from './json-render'
 export * from './rpc'
 export * from './rpc-augments'
 export * from './utils'
